Implement buffered audio, implement multiple cutting strategies

+++ src/config/cli.rs
@@ -5,21 +5,25 @@ use clap::{Parser, Subcommand};
 use super::{Service, SoundServer};
 
 #[derive(Parser, Debug, Clone)]
-#[command(version, about, long_about = None)]
+#[command(version, about)]
+#[command(propagate_version = true)]
 pub struct CliOpts {
     /// The output directory to record to.
     /// Passing this argument will override the setting
     /// in the config file.
+    #[arg(long)]
     pub output_dir: Option<PathBuf>,
     /// The service to record.  Passing this argument will override
     /// the setting in the config file.
+    #[arg(long)]
     pub service: Option<Service>,
     /// The sound server to use.  Passing this argument will override
     /// the setting in the config file.
+    #[arg(long)]
     pub sound_server: Option<SoundServer>,
     #[arg(short, long, action = clap::ArgAction::Count)]
-    pub verbosity: usize,
-    #[clap(subcommand)]
+    pub verbosity: u8,
+    #[command(subcommand)]
     pub command: Command,
 }
 
@@ -35,7 +39,6 @@ pub enum Command {
 }
 
 #[derive(Parser, Debug, Clone)]
-#[command(version, about, long_about = None)]
 pub struct CutArgs {
     /// Path of the recording session to cut.
     pub path: PathBuf,
